Fixed valgrind errors with ObjectModelBase and ObjectModelDelegate
Valgrind api could not be directly be used in getConsumedSizeInBytesWithHeader in ObjectModelBase since it requires size of header/object and
this method is to find size itself.
So we need to go down to ObjectModelDelegate in example and implement it there.

Also fixed code acc. to code review

As per Andrew's review I have -
1. Added #define check while defining valgrindMemppolAddr
2. Fixed the accidental line switch

Issue: #826
Signed-off-by: Varun Garg <varun.10@live.com>