improved valgrind calls in addFreeMemory
Remaining valgrind issue is with the calls from final else condition - "/* Check if the hole is a free list candidate */ ".
If we skip valgrind call in that block, we get errors of a few overwritten objects.
The overlapping errors are still less than before, meaning the other valgrind calls are ok in the class.
If we add the call, we get errors due to inaccessable slot objects later.

Issue: #826
Signed-off-by: Varun Garg <varun.10@live.com>